+++ R/predict_raster.R
@@ -2,10 +2,17 @@
 #'
 #' This function allows to use a raster as data to make predictions from a
 #' variety of `tidymodels` objects, such as [`simple_ensemble`] or
-#' [`stacks::stacks`]
+#' `stacks::stacks`
 #' @param object the `tidymodels` object of interest
 #' @param raster the [`terra::SpatRaster`] or `stars` with the input data. It
 #'   has to include levels with the same names as the variables used in `object`
+#' @param filename the name of the output file raster file; this is only needed
+#'   to save large rasters, and can be left blank for rasters that can be kept
+#'   in memory.
+#' @param n positive integer indicating how many copies the data may be in
+#'   memory at any point in time (it defaults to 4). This is used to determine
+#'   whether rasters can be processed in one go, or in chunks. If you get an out
+#'   of memory error, increase `n`. See [terra::writeStart()] for more details.
 #' @param ... parameters to be passed to the standard `predict()` function for
 #'   the appropriate object type (e.g. `metric_thresh` or `class_thresh`).
 #' @returns a [`terra::SpatRaster`] (or `stars` if that is the input) with the
@@ -20,51 +27,109 @@ predict_raster <- function(object, raster, ...) {
 
 #' @rdname predict_raster
 #' @export
-predict_raster.default <- function(object, raster, ...) {
+predict_raster.default <- function(object, raster, filename = "", n = 4,
+                                   ...) {
   if (inherits(raster, "stars")) {
     is_stars <- TRUE
     raster <- as(raster, "SpatRaster")
   } else {
     is_stars <- FALSE
   }
-  # create a dataframe
-  raster_df <- terra::as.data.frame(raster, cell = TRUE, na.rm = TRUE)
-  # create a vector of predictions by dispatching to the predict generics
-  pred_df <- stats::predict(object, raster_df, ...)
 
-  # create an empty raster where to put the predictions (using the original
-  # raster as a template)
-  pred_raster <- terra::rast(raster[[1]])
+  # we need to figure out how many layers we will need in the output raster
+  # we predict with 5 random points
+  rast_sub_values <- terra::spatSample(raster,
+    size = 5,
+    na.rm = TRUE,
+    replace = FALSE,
+    method = "random"
+  )
 
-  # and now fill in the values, adding a layer for each aggregating function we
-  # used
-  pred_raster[raster_df$cell] <- pred_df %>% dplyr::pull(1)
-  if (is.factor(pred_df %>% dplyr::pull(1))) {
-    # nolint start
-    # old code, kept for reference
-    # levels(pred_raster) <-
-    #          data.frame(id = 1:2, class = levels(pred_df %>% dplyr::pull(1)))
-    # nolint end
-    # make predict_raster work with multilevel predictions
-    # edit by @piabenaud
-    levels(pred_raster) <- data.frame(
-      id = seq_along(dplyr::pull(unique(pred_df))),
-      class = levels(pred_df %>% dplyr::pull(1))
+  if (nrow(rast_sub_values) == 0) {
+    stop("We could not find any non-NA value in the raster")
+  }
+
+  # make predictions
+  pred <- stats::predict(object, rast_sub_values, ...)
+  n_layers_out <- ncol(pred)
+  layer_names <- names(pred)
+  rm(pred)
+
+
+  # start reading the raster
+  terra::readStart(raster)
+  on.exit(terra::readStop(raster))
+
+  # now create an output raster with the correct number of layers
+  pred_raster <- terra::rast(raster, nlyr = n_layers_out)
+  pred_raster_out <- terra::writeStart(pred_raster, filename = filename, n = n)
+
+
+  for (i in 1:pred_raster_out$n) {
+    # read the values
+    rast_sub_values <- terra::readValues(raster,
+      pred_raster_out$row[i],
+      pred_raster_out$nrows[i],
+      dataframe = TRUE
+    )
+    tot_df_rows <- nrow(rast_sub_values)
+    # add row numbers to the data frame
+    rast_sub_values <- rast_sub_values %>%
+      dplyr::mutate(row = dplyr::row_number()) %>%
+      dplyr::filter(stats::complete.cases(rast_sub_values))
+    # remove lines with any NA
+    # this is important, as the predict function will not work with NA values
+    # and we need to remove them before passing the data to the predict function
+
+    # create a data.frame with predictions that has the same number of rows
+    # as the original data frame
+    pred_all <- data.frame(matrix(NA_real_,
+      nrow = tot_df_rows,
+      ncol = n_layers_out
+    ))
+    names(pred_all) <- layer_names
+
+    # make predictions (only if we have some values to predict)
+    if (nrow(rast_sub_values) > 0) {
+      pred <- stats::predict(object, rast_sub_values, ...)
+      pred_all[rast_sub_values$row, ] <- pred
+    }
+
+
+    # write the values
+    terra::writeValues(
+      pred_raster, as.matrix(pred_all),
+      pred_raster_out$row[i],
+      pred_raster_out$nrows[i]
     )
   }
 
+  for (i in seq_len(ncol(pred))) {
+    # if a given prediction is a factor, we need to convert the relevant layer
+    if (is.factor(pred %>% dplyr::pull(i))) {
+      levels_in_factor <- levels(pred %>% dplyr::pull(i))
 
-  if (ncol(pred_df) > 1) {
-    for (i_col in 2:ncol(pred_df)) {
-      pred_raster2 <- terra::rast(raster[[1]])
-      pred_raster2[raster_df$cell] <- pred_df %>% dplyr::pull(i_col)
-      terra::add(pred_raster) <- pred_raster2
+      levels(pred_raster[[i]]) <-
+        data.frame(
+          id = seq_along(levels_in_factor),
+          class = levels_in_factor
+        )
     }
   }
-  names(pred_raster) <- names(pred_df)
-  if (is.factor(pred_df %>% dplyr::pull(1))) {
+
+  # update names of the prediction raster
+  names(pred_raster) <- layer_names
+  # update if it is a factor
+  if (is.factor(pred %>% dplyr::pull(1))) {
     names(pred_raster) <- paste0("binary_", names(pred_raster))
   }
+
+  # set the time to match the raster of origin
+  terra::time(pred_raster, tstep = terra::timeInfo(raster)$step) <-
+    rep(terra::time(raster)[1], terra::nlyr(pred_raster))
+
+  terra::writeStop(pred_raster)
+
   if (is_stars) {
     pred_raster <- stars::st_as_stars(pred_raster, as_attributes = TRUE)
   }